src/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.6.4
psycopg2-binary==2.9.9
